move getTranslationPaths to commoon
